Makes several changes to landing page.
This commit makes multiple changes to the landing page, including
adjusting the font size of elements on larger screens, turning the
first part of the landing page into a component, and fixing the header
to the top of the page. The header's text and background colors also
adjust based on if the user has scrolled or not.